+++ test/cli.js
@@ -24,7 +24,10 @@ describe('cli', function () {
         fs: bag.mock.fs(checks, mocks),
         './repoman': function(config, scms) {
           return {
-            run: function (command, exit) {
+            config: function (exit) {
+              checks.repoman_config_exit = exit;
+            },
+            exec: function (command, exit) {
               checks.repoman_run_command = command;
               checks.repoman_run_exit = exit;
             }
@@ -57,35 +60,41 @@ describe('cli', function () {
       checks.bag_parse_dir.should.equal('/somedir/repoman/lib');
     });
 
-    it('should contain delete commands and deletegate to repoman run when exec is called', function () {
+    it('should contain config commands and delegate to repoman config when exec is called', function () {
+      checks.bag_parse_commands['config'].desc.should.equal('create sample configuration file');
+      checks.bag_parse_commands['config'].action();
+      checks.repoman_config_exit.should.be.a('function');
+    });
+
+    it('should contain delete commands and delegate to repoman exec when exec is called', function () {
       checks.bag_parse_commands['delete'].desc.should.equal('delete the local repositories');
       checks.bag_parse_commands['delete'].action();
       checks.repoman_run_command.should.equal('delete');
       checks.repoman_run_exit.should.be.a('function');
     });
 
-    it('should contain init commands and deletegate to repoman run when exec is called', function () {
+    it('should contain init commands and delegate to repoman exec when exec is called', function () {
       checks.bag_parse_commands.init.desc.should.equal('initialise the repositories');
       checks.bag_parse_commands.init.action();
       checks.repoman_run_command.should.equal('init');
       checks.repoman_run_exit.should.be.a('function');
     });
 
-    it('should contain get commands and deletegate to repoman run when exec is called', function () {
+    it('should contain get commands and delegate to repoman exec when exec is called', function () {
       checks.bag_parse_commands.get.desc.should.equal('update local repositories with changes from remote repositories');
       checks.bag_parse_commands.get.action();
       checks.repoman_run_command.should.equal('get');
       checks.repoman_run_exit.should.be.a('function');
     });
 
-    it('should contain changes commands and deletegate to repoman run when exec is called', function () {
+    it('should contain changes commands and delegate to repoman exec when exec is called', function () {
       checks.bag_parse_commands.changes.desc.should.equal('display the changes in local repositories');
       checks.bag_parse_commands.changes.action();
       checks.repoman_run_command.should.equal('changes');
       checks.repoman_run_exit.should.be.a('function');
     });
 
-    it('should contain save commands and deletegate to repoman run when exec is called', function () {
+    it('should contain save commands and delegate to repoman exec when exec is called', function () {
       checks.bag_parse_commands.save.desc.should.equal('update remote repositories with changes from local repositories');
       checks.bag_parse_commands.save.action();
       checks.repoman_run_command.should.equal('save');